在线比武配置

+++ src/main/webapp/js/competition/pkOnlSetting.js
@@ -18,14 +18,18 @@ var vm = new Vue({
             totalCount: 0
         },
         rules: {//表单验证规则
-            // smallNum: [
-            //     {required: true, message: '请输入参数名', trigger: 'blur'},
-            //     {max: 3, message: '最大长度3', trigger: 'blur'}
-            // ],
-            // code: [
-            //     {required: true, message: '请输入参数值', trigger: 'blur'},
-            //     {max: 50, message: '最大长度50', trigger: 'blur'}
-            // ]
+             topicNum: [
+                 {required: true, message: '请选择', trigger: 'blur'},
+             ],
+            uniformRules: [
+                {required: true, message: '请选择', trigger: 'blur'},
+
+            ],
+            answerTime: [
+                {required: true, message: '请选择', trigger: 'blur'},
+
+            ],
+
         },
         dialogConfig: false,//table弹出框可见性
         dialog2:false,//查看小关详情弹出框
@@ -63,37 +67,45 @@ var vm = new Vue({
             this.reload();
         },
         // 保存和修改
-        saveOrUpdate: function () {
+        saveOrUpdate: function (formName) {
+            this.$refs[formName].validate(function (valid) {
 
-            //保存前默认先删除一波
-            $.ajax({
-                type: "POST",
-                url: baseURL + 'competitionOnline/deleteAll',
-                async:false,
-                dataType: "json",
-                success: function (result) {
-                    var url ="competitionOnline/save";
-                    $.ajax({
-                        type: "POST",
-                        url: baseURL + url,
-                        contentType: "application/json",
-                        data: JSON.stringify(vm.competitionOnline),
-                        success: function (result) {
-                            if (result.code === 0) {
-                                vm.$alert('操作成功', '提示', {
-                                    confirmButtonText: '确定',
-                                    callback: function () {
-                                        vm.dialogConfig = false;
-                                        vm.reload();
+                    if (valid) {
+                        //保存前默认先删除一波
+                        $.ajax({
+                            type: "POST",
+                            url: baseURL + 'competitionOnline/deleteAll',
+                            async:false,
+                            dataType: "json",
+                            success: function (result) {
+                                var url ="competitionOnline/save";
+                                $.ajax({
+                                    type: "POST",
+                                    url: baseURL + url,
+                                    contentType: "application/json",
+                                    data: JSON.stringify(vm.competitionOnline),
+                                    success: function (result) {
+                                        if (result.code === 0) {
+                                            vm.$alert('操作成功', '提示', {
+                                                confirmButtonText: '确定',
+                                                callback: function () {
+                                                    vm.dialogConfig = false;
+                                                    vm.reload();
+                                                }
+                                            });
+                                        } else {
+                                            alert(result.msg);
+                                        }
                                     }
                                 });
-                            } else {
-                                alert(result.msg);
                             }
-                        }
-                    });
-                }
+                        });
+                    }
+                    else {
+                        return false;
+                    }
             });
+
 
 
         },
@@ -155,6 +167,7 @@ var vm = new Vue({
                 data: {type:"QUESTION_TYPE",Parentcode:"0"},
                 success: function (result) {
                     // vm.itemtype=result.dictlist;
+                    vm.itemtype=[];
                     for(var tt=0;tt<result.dictlist.length;tt++)
                     {
                         if((result.dictlist[tt].value=="单选题")||(result.dictlist[tt].value=="判断题") )
@@ -185,7 +198,7 @@ var vm = new Vue({
                     if(result.page.list.length!="0")
                     {
 
-                        vm.$alert('请先删除原有配置，在添加新的配置');
+                        vm.$alert('请先删除原有配置，再添加新的配置');
                         vm.dialogConfig = false;
                     }
                     else
@@ -210,38 +223,39 @@ var vm = new Vue({
         },
         onselectuniformRules:function (num) {//点完是否统一配置触发事件
 
-          var topicnum=  Number(num.topicNum);  //题目数量
+                  var topicnum=  Number(num.topicNum);  //题目数量
 
-          var uniformRules=num.uniformRules  //是否统一配置
+                  var uniformRules=num.uniformRules  //是否统一配置
 
-          if(!topicnum)
-          {
-              vm.$message({
-                  type: 'info',
-                  message: '请先设置题目数量!'
-              });
-              return;
-          }
-          //不管选的是是还是否  都先清除一遍  在加
-            num.battleTopicSettingList=[];
-          if(uniformRules=="0")//0不是统一配置
-          {
-                for(var p=0;p<topicnum;p++)
-                {
-                    num.battleTopicSettingList.push
-                    (
-                        {id:'',howManySmall:p+1,}
-                    )
-                }
+                  if(!topicnum)
+                  {
+                      vm.competitionOnline.uniformRules='';
+                      vm.$message({
+                          type: 'info',
+                          message: '请先设置题目数量!'
+                      });
+                      return;
+                  }
+                  //不管选的是是还是否  都先清除一遍  在加
+                    num.battleTopicSettingList=[];
+                  if(uniformRules=="0")//0不是统一配置
+                  {
+                        for(var p=0;p<topicnum;p++)
+                        {
+                            num.battleTopicSettingList.push
+                            (
+                                {id:'',howManySmall:p+1,}
+                            )
+                        }
 
-          }
-          else if(uniformRules=="1")//1是  是统一
-          {
-              num.battleTopicSettingList.push
-              (
-                  {id:'',}
-              )
-          }
+                  }
+                  else if(uniformRules=="1")//1是  是统一
+                  {
+                      num.battleTopicSettingList.push
+                      (
+                          {id:'',}
+                      )
+                  }
 
         },
         look: function (index, row) {
@@ -267,7 +281,13 @@ var vm = new Vue({
             });
         },
         del: function () {
-
+            if(vm.tableData.length=='0'){
+                vm.$message({
+                    type: 'info',
+                    message: '占无数据!'
+                });
+                return;
+            }
             this.$confirm('此操作将永久删除在线比武配置, 是否继续?', '提示', {
                 confirmButtonText: '确定',
                 cancelButtonText: '取消',
@@ -296,6 +316,14 @@ var vm = new Vue({
 
         },
         update: function () {
+
+            if(vm.tableData.length=='0'){
+                vm.$message({
+                    type: 'info',
+                    message: '占无数据!'
+                });
+                return;
+            }
             // 每次进来先制空
             vm.checkNum=[];
             vm.competitionOnline=[];
